==> 0. Loading of Emissions Data.ipynb <==
# Loading Emissions Data

#### Imports

# Basic imports
import pandas as pd
import xarray as xr
import os

#### Index

1. [Introduction](#intro)
2. [CH$_4$ Data](#ch4)
3. [CO$_2$ Date](#co2)
4. [N$_2$O](#n2o)

---
<a id='intro'></a>
## Introduction

Given the link between global warming and wildfires, a certain measurable feature that we can investigate is the amount of greenhouse gases produced for a given longitude and latitude. Such precision requires a data set that is extremely large and in our case is spread across files for individual years. In order to make this information more accessible, we will combine the currently separated files into a single `.csv` file that can be read into the Jupyter Notebook. This notebook therefore describes the process of reducing the size of the emissions data that we are going to use in our project, showing how we removed redundant values and created multiple `.csv` files that we may access later on in our project.

---
<a id='ch4'></a>
## CH$_4$ Data

Before we begin extracting the relevant columns, we need to understand the structure of the data. The following dataset, as well as all subsequent datasets pertaining to atmospheric greenhouse gas emissions is collected by the Emissions Database for Global Atmospheric Research (EDGAR), spearheaded by the European Commission. The data in all the files downloaded is organised in the same way, meaning that we will only describe the layout once within this section, as this will also apply to the other instances (CO$_2$ and N$_2$O). 

The data is gridded and stored in a .nc file. The grid is measured by $0.1^\circ × 0.1^\circ$ change in longitude and latitude. The data is split into separate .nc files for individual years, meaning that for each .nc file, the following information is shown:

|Column|Description|
|:--|:--|
|lat|Latitude|
|lon|Longitude|
|emi_*substance*|The amount of the respective substance released. Measured in kg substance/m$^2$|

We will read the files for the individual years, from which we will create a DataFrame for the given gas. These can hopefully be used and read into our notebooks later on in the project when they are required.

# Create DataFrame
ch4_df = pd.DataFrame()

# Iterate through the files in the directory
for file in os.listdir('EDGAR Emissions/CH4_TOTALS_nc'):
    if file[-2:] == 'nc':
        # Get the year of the file
        year = file[9:13]
        path = f'EDGAR Emissions/CH4_TOTALS_nc/{file}'
        ds = xr.open_dataset(path)
        tmp_df = ds.to_dataframe()
        tmp_df.unstack()
        # Add year column to DataFrame
        tmp_df['year'] = int(year)
        # Concatenate the file into a single df
        ch4_df = pd.concat([ch4_df, tmp_df])

Ultimately, we have created a DataFrame, that has a multiindex. Even after unstacking, the `lat` and `lon` index columns are not accessible. Consequently, I saved the DataFrame to a .csv file that we will be able to read from later on. 

ch4_df.to_csv('data/EDGAR Emissions/ch4.csv')

ch4_df.head()

If we read the csv into a DataFrame, we will see how the data is laid out.

ch4_df = pd.read_csv('EDGAR Emissions/ch4.csv')

ch4_df.head()

In order to use this data productively, we need to ensure that the `lat` and `lon` values are written in a way that is consistent with our wildfire data.

print('The longitude values range from:')
print(ch4_df['lon'].min(), ch4_df['lon'].max())

print('The latitude values range from:')
print(ch4_df['lat'].min(), ch4_df['lat'].max())

This means that the longitude values are all positive, while the latitude values range from -90$^{\circ}$ to 90$^{\circ}$. The longitude values that we have defined for the area of North America are defined in a range from -180$^{\circ}$ to 180$^{\circ}$. We need to convert these values so that they are in accordance with the DataFrame.

|Westernmost Longitude| Easternmost Longitude|Northenmost Latitude|Southernmost Latitude|
|:--|:--|:--|:--|
|-178.1333 /  181.8667|-53.0567 / 306.9433|82.9143|14.0749|

From these values we can filter out the values that don't fall within these geographical boundaries.

latitudes = (ch4_df['lat'] >= 14.0749) & (ch4_df['lat'] <= 82.9143)
longitudes = (ch4_df['lon'] >= 181.8667) & (ch4_df['lon'] <= 306.9433)
ch4_df = ch4_df[latitudes & longitudes]

Now that we have filtered out the values that we want we have a DataFrame with relevant latitudes and longitudes for the years 1992-2015.

ch4_df.head()

We notice however that the DataFrame is not sorted. Perhaps it would be best if we sort this by year for easier reading.

ch4_sorted = ch4_df.sort_values(by='year')
ch4_sorted.head()

ch4_sorted.to_csv('EDGAR Emissions/ch4.csv')

---
<a id='co2'></a>
## $CO_2$ Data

Considering the similarity between this dataset and the previous, we will follow the same process as before to create a similar `.csv` file.

co2_df = pd.DataFrame()

for file in os.listdir('EDGAR Emissions/CO2_excl_short-cycle_TOTALS_nc'):
    if file[-2:] == 'nc':
        year = file[-22:-18]
        path = f'EDGAR Emissions/CO2_excl_short-cycle_TOTALS_nc/{file}'
        ds = xr.open_dataset(path)
        tmp_df = ds.to_dataframe()
        tmp_df.unstack()
        tmp_df['year'] = int(year)
        co2_df = pd.concat([co2_df, tmp_df])
    
co2_df.head()

co2_df.to_csv('EDGAR Emissions/co2.csv')

co2_df = pd.read_csv('EDGAR Emissions/co2.csv')

co2_df.head()

co2_df.shape

co2_df.isnull().sum()

print('The longitude values range from:')
print(co2_df['lon'].min(), co2_df['lon'].max())

print('The latitude values range from:')
print(co2_df['lat'].min(), co2_df['lat'].max())

Find the rows with the appropriate latitude and longitude.

latitudes = (co2_df['lat'] >= 14.0749) & (co2_df['lat'] <= 82.9143)
longitudes = (co2_df['lon'] >= 181.8667) & (co2_df['lon'] <= 306.9433)
co2_df = co2_df[latitudes & longitudes]

co2_df.shape

co2_sorted = co2_df.sort_values(by='year')
co2_sorted.head()

co2_sorted.to_csv('data/EDGAR Emissions/co2.csv')

---
<a id='n2o'></a>
## $N_2O$ Data

n2o_df = pd.DataFrame()

for file in os.listdir('EDGAR Emissions/N2O_TOTALS_nc'):
    if file[-2:] == 'nc':
        year = file[9:13]
        path = f'EDGAR Emissions/N2O_TOTALS_nc/{file}'
        ds = xr.open_dataset(path)
        tmp_df = ds.to_dataframe()
        tmp_df.unstack()
        tmp_df['year'] = int(year)
        n2o_df = pd.concat([n2o_df, tmp_df])

n2o_df.head()

n2o_df.shape

n2o_df.to_csv('EDGAR Emissions/n2o.csv')

n2o_df = pd.read_csv('EDGAR Emissions/n2o.csv')

n2o_df.head()

n2o_df.shape

n2o_df.isnull().sum()

print('The longitude values range from:')
print(n2o_df['lon'].min(), n2o_df['lon'].max())

print('The latitude values range from:')
print(n2o_df['lat'].min(), n2o_df['lat'].max())

latitudes = (n2o_df['lat'] >= 14.0749) & (n2o_df['lat'] <= 82.9143)
longitudes = (n2o_df['lon'] >= 181.8667) & (n2o_df['lon'] <= 306.9433)
n2o_df = n2o_df[latitudes & longitudes]
n2o_df.head()

n2o_df.shape

n2o_df = n2o_df.sort_values(by='year')
n2o_df.head()

n2o_df.to_csv('EDGAR Emissions/n2o.csv')
