==> 1.1. Gathering Weather Data.ipynb <==
# Gathering Weather Data

#### Imports

# Basic imports
import pandas as pd
import numpy as np
import urllib.request
import json
import sys
from ipywidgets import IntProgress
from IPython.display import display

# Custom imports
import api 
import utils

#### Index

1. [Introduction](#intro)
2. [Loading Data](#ld)
3. [API Call](#api)
4. [Saving Data](#saving)

---
<a id='intro'></a>
## Introduction

This notebook is dedicated entirely to the collection of weather data given the sample that we have created. The way in which we will collect the weather data is through an API call to <a href='https://www.visualcrossing.com/'>*Visual Crossing*</a>, a weather api which provides detailed weather information given a specific longitude and latitude. The API call will collect the data available for the 7 days prior to the start of the wildfire, storing these values in a list so we may access them at a later stage, whilst simultaneously calculating the average of these values. The weather information that we will collect will be the following:

<table>
  <tr>
    <th style="text-align: left; background: lightgrey">Weather Metric</th>
    <th style="text-align: left; background: lightgrey">Description</th>
    <th style="text-align: left; background: lightgrey">Unit</th>
  </tr>
  <tr>
    <td style="text-align: left"> <code>tempmax</code> </td>
    <td style="text-align: left">Maximum Temperature</td>
    <td style="text-align: left">Degrees Celcius</td>
  </tr>
  <tr>
    <td style="text-align: left"><code>temp</code></td>
    <td style="text-align: left">Temperature (or mean temperature)</td>
    <td style="text-align: left">Degrees Celcius</td>
  </tr>
  <tr>
    <td style="text-align: left"><code>humidity</code></td>
    <td style="text-align: left">Relative Humidity</td>
    <td style="text-align: left">%</td>
  </tr>
  <tr>
    <td style="text-align: left"><code>precip</code></td>
    <td style="text-align: left">Precipitation</td>
    <td style="text-align: left">Millimetres</td>
  </tr>
  <tr>
    <td style="text-align: left"><code>dew</code></td>
    <td style="text-align: left">Dew Point</td>
    <td style="text-align: left">Degrees Celsius</td>
  </tr>
  <tr>
    <td style="text-align: left"><code>windspeed</code></td>
    <td style="text-align: left">Wind Speed</td>
    <td style="text-align: left">Kilometres Per Hour</td>
  </tr>
  <tr>
    <td style="text-align: left"><code>winddir</code></td>
    <td style="text-align: left">Wind Direction</td>
    <td style="text-align: left">Degrees</td>
  </tr>
  <tr>
    <td style="text-align: left"><code>pressure</code></td>
    <td style="text-align: left">Sea Level Pressue</td>
    <td style="text-align: left">Millibars (Hectopascals)</td>
  </tr>
</table>

Once we have collected the data, we will store it as a `.pkl` file so we may access it again in other sections of the project.

---

<a id='ld'></a>
## Loading Data

As we have created a sample of wildfires previously, we will load this data in and collect the weather data for the longitude and latitude values contained within. We can load the data from the `.pkl` file that we saved in previous notebooks.

# Load the data
df = pd.read_pickle('sample_data/sample.pkl')

# Check the output
df.head()

# Check the shape
df.shape

---
<a id='api'></a>
## API Call

The API call that we are making is going to collect the weather information for a given longitude, latitude, and for a range of dates. Given the date of a wildfire, we are going to collect the weather information for the week prior. We are going to save the weather data for the past 7 days as a list, for each individual metric, and calculate the average. All of the data that is being collected is being saved into a dictionary that, once all the rows in the sample DataFrame have been parsed, will be converted into a DataFrame and concatenated into a unified DataFrame which will now have the wildfire data, alongside the relevant weather data.

A few questions arise regarding the API call. The first may be why is the weather data for the week being stored in the first place; why are we not calculating important variables from this data directly? We are storing the information from the past 7 days so that we may access this in case it is required during later stages of our project. Although we have already thought of what variables might be interesting to calculate from the historical weather data (variance and the change in the metric from 1 week prior to the day of the wildfire), there may be other metrics which we can calculate but have failed to recognise so far. 

A second question pertains to why we are only collecting the weather information for the past 7 days. The answer to thiis question has multiple dimensions, but at its core 7 days seems to be the best balance for maximising speed and minimising the cost to run the API call.

Having discussed some of the reasoning behind the API call, we can implement it, using the following code:

# Get API Key
key = api.API_KEY

# BaseURL
BaseURL = 'https://weather.visualcrossing.com/VisualCrossingWebServices/rest/services/timeline/'

# Data dictionary into which the weather data will be appended
data = {
    'tempmax': [],
    'avg_tempmax': [],
    'temp': [], 
    'avg_temp': [],
    'humidity': [],
    'avg_humidity': [],
    'precip': [], 
    'avg_precip':[],
    'dew': [], 
    'avg_dew' :[],
    'windspeed': [],
    'avg_windspeed': [],
    'winddir': [],
    'avg_winddir': [],
    'pressure': [],
    'avg_pressure': []
}

# Instantiate the bar
progress = IntProgress(min=0, max=df.shape[0]) 

# Display the bar
display(progress) 

# Iterate through the dataframe and calculate weather data
for index, row in df.iterrows():
    
    # Update progress bar
    progress.value += 1
    
    # Create the table variables
    end_date = row['DATE']
    start_date = end_date - pd.Timedelta(6, 'days')
    latitude = row['LATITUDE']
    longitude = row['LONGITUDE']
    
    # Create the API Query
    query = f'{latitude}%2C%20{longitude}/{start_date.date()}/{end_date.date()}?unitGroup=metric&include=days&key=XYAH73VX8WMHKJ3YE6NG62R6V&contentType=json'
    url = BaseURL + query
    
    try: 
        response = urllib.request.urlopen(url)
        # Parse the results as JSON
        string = response.read().decode('utf-8')
        jsonData = json.loads(string)
    except urllib.error.HTTPError  as e:
        ErrorInfo= e.read().decode() 
        print('Error code: ', e.code, ErrorInfo)
        sys.exit()

    # Create lists for the values of the last 7 days
    variables = {
        'tempmax': [],    
        'temp': [],
        'humidity': [],
        'precip': [],
        'dew': [],
        'windspeed': [],
        'winddir': [],
        'pressure': []
    }
    
    # Parse the json data
    for daily_data in jsonData['days']:
        # Iterate through the variables to get the data for the last 7 days
        for variable in variables:
            if daily_data[variable] is not None:
                variables[variable].append(daily_data[variable]) 
            else:
                variables[variable].append(np.nan)
    
    # Append all the variables to the data dictionary
    for key in data:
        if 'avg_' in key:
            metric = key[4:]
            data[key].append(utils.mean(variables[metric]))
        else:
            data[key].append(variables[key])

---
<a id='saving'></a>
## Saving Data

Now that we have loaded the data into a dictionary, we can append this to our `df` and save this as a `.pkl`

# Convert dictionary into DataFrame
weather_info = pd.DataFrame(data)

weather_info.head()

weather_info.shape

# Concatenate
df_with_weather = pd.concat([df, weather_info], axis=1)

# Check that correctly concatenated
df_with_weather.head()

df_with_weather.shape

# Save the DataFrame as a .pkl
df_with_weather.to_pickle('sample_data/30k_samples_with_weather.pkl')
