==> 1.2. Gathering Emissions Data.ipynb <==
# Gathering Emission Data

As with our *Gathering Weather Data* notebook, this notebook will exclusively focus on collecting the relevant emissions data for a given latitude and longitude. Unlike our other notebook, however, we will not be using an API to collect this data, but will use a data set collected from <a href='https://edgar.jrc.ec.europa.eu/index.php/dataset_ghg60#p2'>EDGAR</a> (Emissions Database for Global Atmospheric Research) which we have aggregated into large CSV files for each of the greenhouse gases being measured; using the data provided by EDGAR, we are able to collect emissions data for $CO_2$ (carbon dioxide), $CH_4$ (methane), and $N_2O$ (nitrous oxide). 

To begin gathering the emission data, we need to understand what our objective is. The data for the emissions is gridded, meaning that the emissions for a specific longitude and latitude along a $0.1^{\circ} \times 0.1^{\circ}$ are measured. Our wildfires have a specific longitude and latitude too, meaning that we are able to find the nearest *node* from the emissions data to be able to derive an approximation for the wildfire latitude and longitude. Our task therefore will be to find the closest node to the latitude and longitude of the wildfire.

---

### Imports

import numpy as np
import pandas as pd

# Custom imports
import emissions_utils
from importlib import reload

---

### Loading Wildfire Sample and CH4 Data

ch4_data = pd.read_csv('EDGAR Emissions/ch4.csv')
wildfires = pd.read_pickle('sample_data/30k_samples_with_weather.pkl')

ch4_data.head()

wildfires.head(2)

# Get unique latitutde, longitude combinations
emissions_latlon = ch4_data[ch4_data['year'] == 1992][['lat', 'lon']]

# Look at the data
emissions_latlon.head()

Check the format of the wildfires latitude and longitude values, as these could have different ranges to the latitude and longitude values of the emissions data. Latitude values can be between -90 and 90, whilst longitude values can be between -180 and 180.

wildfires[['LATITUDE', 'LONGITUDE']].describe()

We see that the latitude and longitude values fall within the ranges mentioned above.

emissions_latlon.describe()

The latitude falls within the range, however the longitude value ranges from 181 to 306, meaning that we need to subtract 360 fromt these values to get longitude values that coincide with our wildfire data.

emissions_latlon['lon'] = emissions_latlon['lon'].apply(lambda x: x - 360)

emissions_latlon.describe()

# Create numpy array of the values
lat_lon = np.array(emissions_latlon)

# Convert the longitude columns of the ch4_data DataFrame
ch4_data['lon'] = ch4_data['lon'].apply(lambda x: x - 360)

ch4_data[['lon']].describe()

# Convert the ch4 into a numpy array
ch4_array = np.array(ch4_data.iloc[:,1:])

# Create list for the emission
ch4_emissions = []

# Iterate through the rows in the wildfires data frame
for _, row in wildfires.iterrows():
    # Get the coordinates of the row
    coordinates = np.array([row['LATITUDE'], row['LONGITUDE']])
    lat, lon = emissions_utils.GetBestLatLon(coordinates, lat_lon)
    index = np.where((ch4_array[:,0] == lat) & (ch4_array[:,1] == lon) & (ch4_array[:,-1] == row['FIRE_YEAR']))
    ch4_emissions.append(ch4_array[index][0][2])

wildfires_with_emissions = wildfires.copy()
wildfires_with_emissions['ch4'] = ch4_emissions

wildfires_with_emissions.head()

### Appending $CO_2$ and $N_2O$ Data

Now that we see how this has been achieved we can automate this process and append the values that we generated.

# Load in the different dataframes
co2_df = pd.read_csv('EDGAR Emissions/co2.csv')
n2o_df = pd.read_csv('EDGAR Emissions/n2o.csv')

# Drop the unnamed column
for df in [co2_df, n2o_df]:
    df.drop('Unnamed: 0', axis=1, inplace = True)

co2_data = emissions_utils.GetEmissions(wildfires, co2_df)

co2_data_df = pd.DataFrame(co2_data)
co2_data_df.head()

co2_data_df.isna().sum()

n2o_data = emissions_utils.GetEmissions(wildfires, n2o_df)

n2o_data_df = pd.DataFrame(n2o_data)
n2o_data_df.head()

n2o_data_df.isna().sum()

wildfires_with_emissions['co2'] = co2_data['emission']
wildfires_with_emissions['n2o'] = n2o_data['emission']

wildfires_with_emissions.head()

wildfires_with_emissions.shape

Now that we have appended the gas emissions data to the wildfires DataFrame, we can save the data into a `.pkl` file which we will be able to analyse in later stages.

wildfires_with_emissions.to_pickle('sample_data/30k_wildfires_weather_emissions.pkl')
