==> 2.1. Creating Further Samples.ipynb <==
#### Imports

# Basic imports
import numpy as np
import pandas as pd
import sqlite3

# Custom imports
import utils
import emissions_utils
import preparation_utils

#### Index

1. [Introduction](#intro)<br>
2. [Retrieving Wildfire Data](#rw)<br>
3. [Gathering Weather Data](#gwd)<br>
4. [Adding Emissions](#ae)<br>
5. [Data Cleaning](#dc)<br>
6. [Calculating Additional Features](#caf)<br>

---
<a id='intro'></a>
# Introduction

As we noted previously, our lack of larger fires is damaging to the accuracy and performance of the classification models that we hope to build. While we can use upsampling techniques in order to improve the distribution, the best alternative is to get more data from the actual observations that we have in our wildfire data set. While we will be able to get more observations from the minority classes, the uneven distribution means that overall the total number of observations will not be massively increased.

Regardless, we will continue by removing the fires that we have already collected in previous samples and collecting fires that only belong to fire classes D and above. We will then replicate the process that we carried out previously - collecting weather and emissions data for the newly acquired observations - ultimately concatenating this onto the sample that we initially generated.

---
<a id='rw'></a>
## Retrieving Wildfire Data
To begin, we will have to query the `.sqlite` file that contains the wildfire data.

# Get only the relevant columns
con = sqlite3.connect("wildfire_data/wildfires.sqlite")
query = """
    SELECT FIRE_YEAR, DISCOVERY_DOY, FIRE_SIZE, FIRE_SIZE_CLASS, LATITUDE, LONGITUDE, STATE
    FROM fires
"""
wildfires = pd.read_sql_query(query, con)

# Drop duplicates
wildfires.drop_duplicates(inplace=True)

# Create the DATE column
wildfires['DATE'] = pd.to_datetime(wildfires['FIRE_YEAR'] * 1000 + wildfires['DISCOVERY_DOY'], format='%Y%j')

# Pop the column
date = wildfires.pop('DATE')

# Insert it into the relevant position
wildfires.insert(0, 'DATE', date)

wildfires.head()

As we are only concerned with getting wildfires from classes D, E, F, and G, we will create a temporary DataFrame in which only these classes are present.

g = wildfires[wildfires['FIRE_SIZE_CLASS'].isin(['D', 'E', 'F', 'G'])]['FIRE_SIZE_CLASS']

utils.count_percentage_df(g)

As expected, the number of very large fires (classes F, G) are still in a relative minority. Since we want to improve the distribution of our data - and for considerations of time - we will gather 3773 observations from each class.

classes = ['D', 'E', 'F', 'G']
all_indexes = np.array([])

for fire_class in classes:
    tmp = wildfires[wildfires['FIRE_SIZE_CLASS']==fire_class].sample(3773)
    all_indexes = np.concatenate((all_indexes, tmp.index.values), axis=None)
    
len(all_indexes) == 3773 * 4

As we have the indexes of the observations that we have already sampled, we can remove those indexes that have already appeared in our intial sample.

sample = pd.read_pickle('data/30k_wildfires_weather_emissions.pkl')
already_collected = sample['index'].values

all_indexes = [idx for idx in all_indexes
               if idx not in already_collected]

len(all_indexes)

We see, then, that after finding the wildfires that have already sampled, we are left with 14,863 new observations, which have the following count and distribution:

df = wildfires.loc[all_indexes]

utils.count_percentage_df(df['FIRE_SIZE_CLASS'])

---
<a id='gwd'></a>
## Gathering Weather Data
Having collected the wildfire data it is time to use the API call to gather the weather information for each longitude and latitude.

# Necessary import
import api
import urllib.request
import json
import sys

# Aesthetic imports
from ipywidgets import IntProgress
from IPython.display import display

# Get API Key
key = api.API_KEY

# BaseURL
BaseURL = 'https://weather.visualcrossing.com/VisualCrossingWebServices/rest/services/timeline/'

# Data dictionary into which the weather data will be appended
data = {
    'tempmax': [],
    'avg_tempmax': [],
    'temp': [], 
    'avg_temp': [],
    'humidity': [],
    'avg_humidity': [],
    'precip': [], 
    'avg_precip':[],
    'dew': [], 
    'avg_dew' :[],
    'windspeed': [],
    'avg_windspeed': [],
    'winddir': [],
    'avg_winddir': [],
    'pressure': [],
    'avg_pressure': []
}

# Instantiate the bar
progress = IntProgress(min=0, max=df.shape[0]) 

# Display the bar
display(progress) 

# Iterate through the dataframe and calculate weather data
for index, row in df.iterrows():
    
    # Update progress bar
    progress.value += 1
    
    # Create the table variables
    end_date = row['DATE']
    start_date = end_date - pd.Timedelta(6, 'days')
    latitude = row['LATITUDE']
    longitude = row['LONGITUDE']
    
    # Create the API Query
    query = f'{latitude}%2C%20{longitude}/{start_date.date()}/{end_date.date()}?unitGroup=metric&include=days&key=XYAH73VX8WMHKJ3YE6NG62R6V&contentType=json'
    url = BaseURL + query
    
    try: 
        response = urllib.request.urlopen(url)
        # Parse the results as JSON
        string = response.read().decode('utf-8')
        jsonData = json.loads(string)
    except urllib.error.HTTPError  as e:
        ErrorInfo= e.read().decode() 
        print('Error code: ', e.code, ErrorInfo)
        sys.exit()

    # Create lists for the values of the last 7 days
    variables = {
        'tempmax': [],    
        'temp': [],
        'humidity': [],
        'precip': [],
        'dew': [],
        'windspeed': [],
        'winddir': [],
        'pressure': []
    }
    
    # Parse the json data
    for daily_data in jsonData['days']:
        # Iterate through the variables to get the data for the last 7 days
        for variable in variables:
            if daily_data[variable] is not None:
                variables[variable].append(daily_data[variable]) 
            else:
                variables[variable].append(np.nan)
    
    # Append all the variables to the data dictionary
    for key in data:
        if 'avg_' in key:
            metric = key[4:]
            data[key].append(utils.mean(variables[metric]))
        else:
            data[key].append(variables[key])

Having stored the weather information in a dictionary, we are able to open this in a DataFrame to inspect some of the elements. 

# Convert dictionary into DataFrame
weather_info = pd.DataFrame(data)

weather_info.head()

df_reset = df.reset_index()

Having reset the index we will be able to concatenate this newly created DataFrame to the DataFrame that contains the wildfire data.

df_with_weather = pd.concat([df_reset, weather_info], axis=1)

df_with_weather.head()

We can also save this DataFrame if we ever want to access this data at some point in the future.

df_with_weather.to_pickle('sample_data/large_fire_samples_with_weather.pkl')

---
<a id='ae'></a>
## Adding Emissions

As before, we also need to find the emissions data for the wildfires.

ch4_df = pd.read_csv('data/EDGAR Emissions/ch4.csv')
co2_df = pd.read_csv('data/EDGAR Emissions/co2.csv')
n2o_df = pd.read_csv('data/EDGAR Emissions/n2o.csv')

# Drop unnamed column
dfs = [ch4_df, co2_df, n2o_df]

for df in dfs:
    df.drop('Unnamed: 0', axis=1, inplace = True)

ch4_data = emissions_utils.GetEmissions(df_with_weather, ch4_df)
co2_data = emissions_utils.GetEmissions(df_with_weather, co2_df)
n2o_data = emissions_utils.GetEmissions(df_with_weather, n2o_df)

df_with_weather['ch4'] = ch4_data['emission']
df_with_weather['co2'] = co2_data['emission']
df_with_weather['n2o'] = n2o_data['emission']

df_with_weather.head()

df_with_weather.to_pickle('sample_data/large_fire_samples_with_weather_emissions.pkl')

---
<a id='dc'></a>
## Data Cleaning

This process will replicate the process that we conducted earlier. Wildfires for which insufficient weather data could be collected will be removed.

df_with_weather = pd.read_pickle('sample_data/large_fire_samples_with_weather_emissions.pkl')

df_raw = df_with_weather.copy()

df_raw.head()

df_raw.isna().sum()

remove_indexes = preparation_utils.GetNullListIndex(df_raw)

len(remove_indexes)

df_cleaned = df_raw.drop(index=remove_indexes)

df_cleaned.isna().sum()

utils.count_percentage_df(df_cleaned['FIRE_SIZE_CLASS'])

df_cleaned.head()

---
<a id='caf'></a>
## Calculating Additional Features

df_cleaned = preparation_utils.GenerateFeatures(df_cleaned)

df_cleaned = preparation_utils.RemoveColumns(df_cleaned)

df_cleaned.head()

df_cleaned.to_pickle('sample_data/large_fires_cleaned.pkl')

Now we will combine the files that we have created in another notebook working from there.
